requirements.txt
alpha-vantage==2.1.3
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
flask-heroku==0.1.9
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
iexfinance==0.4.3
itsdangerous==1.1.0
Jinja2==2.11.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.2.0
numpy==1.18.1
pandas==1.0.1
pip-tools==4.5.1
psycopg2-binary==2.8.4
pycparser==2.20
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.25.8
Werkzeug==1.0.0
WTForms==2.2.1
